Move ODL model classes to their own bundle.
This patch also introduces licenses header to all java files.
