+++ extensions/odl/capabilities/pom.xml
@@ -36,6 +36,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.mqnaas.extensions</groupId>
+			<artifactId>odl.model</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.mqnaas.extensions</groupId>
 			<artifactId>odl.client</artifactId>
 			<version>${project.version}</version>
 		</dependency>
