+++ extensions/odl/capabilities/src/main/java/org/mqnaas/extensions/odl/capabilities/flows/IFlowManagement.java
@@ -1,5 +1,26 @@
 package org.mqnaas.extensions.odl.capabilities.flows;
 
+/*
+ * #%L
+ * MQNaaS :: ODL Capabilities
+ * %%
+ * Copyright (C) 2007 - 2015 Fundació Privada i2CAT, Internet i
+ * 			Innovació a Catalunya
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
 import java.util.Collection;
 
 import javax.ws.rs.BadRequestException;
@@ -15,7 +36,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
 import org.mqnaas.core.api.ICapability;
-import org.mqnaas.extensions.odl.client.hellium.flowprogrammer.api.model.FlowConfig;
+import org.mqnaas.extensions.odl.hellium.flowprogrammer.model.FlowConfig;
 
 @Path("/")
 public interface IFlowManagement extends ICapability {
