+++ extensions/odl/client/pom.xml
@@ -27,6 +27,11 @@
 	</properties>
 
 	<dependencies>
+		<dependency>
+			<groupId>org.mqnaas.extensions</groupId>
+			<artifactId>odl.model</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 		<!-- CXF -->
 		<dependency>
 			<groupId>org.apache.cxf</groupId>
