+++ extensions/odl/client/src/test/java/org/mqnaas/extensions/odl/client/OpenDaylightFlowProgrammerNorthboundClientTest.java
@@ -1,5 +1,26 @@
 package org.mqnaas.extensions.odl.client;
 
+/*
+ * #%L
+ * MQNaaS :: ODL Client
+ * %%
+ * Copyright (C) 2007 - 2015 Fundació Privada i2CAT, Internet i
+ * 			Innovació a Catalunya
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
 import java.util.Arrays;
 
 import org.junit.After;
@@ -8,9 +29,9 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.mqnaas.extensions.odl.client.hellium.flowprogrammer.api.IOpenDaylightFlowProgrammerNorthbound;
-import org.mqnaas.extensions.odl.client.hellium.flowprogrammer.api.model.FlowConfig;
-import org.mqnaas.extensions.odl.client.hellium.flowprogrammer.api.model.FlowConfigs;
-import org.mqnaas.extensions.odl.client.hellium.flowprogrammer.api.model.Node;
+import org.mqnaas.extensions.odl.hellium.flowprogrammer.model.FlowConfig;
+import org.mqnaas.extensions.odl.hellium.flowprogrammer.model.FlowConfigs;
+import org.mqnaas.extensions.odl.hellium.flowprogrammer.model.Node;
