+++ extensions/odl/model/src/main/java/org/mqnaas/extensions/odl/hellium/flowprogrammer/model/ActionType.java
@@ -10,7 +10,28 @@
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
 // package org.opendaylight.controller.sal.action;
-package org.mqnaas.extensions.odl.client;
+
+/*
+ * #%L
+ * MQNaaS :: ODL Model
+ * %%
+ * Copyright (C) 2007 - 2015 Fundació Privada i2CAT, Internet i
+ * 			Innovació a Catalunya
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+package org.mqnaas.extensions.odl.hellium.flowprogrammer.model;
 
 /**
  * The enumeration of actions supported by the controller
