+++ extensions/odl/pom.xml
@@ -16,6 +16,7 @@
 	<description>MQNaaS OpenDaylight</description>
 
 	<modules>
+		<module>model</module>
 		<module>client</module>
 		<module>clientprovider</module>
 		<module>capabilities</module>
